+++ packages/api-graphql/__tests__/APIClient.test.ts
@@ -294,6 +294,32 @@ describe('flattenItems', () => {
 
 			expect(selSet).toEqual(expected);
 		});
+
+		test('mix of related and non-related fields in a nested model creates a nested object with all necessary fields', () => {
+			const selSet = customSelectionSetToIR(
+				modelIntroSchema.models,
+				'CommunityPost',
+				['poll.question', 'poll.answers.answer', 'poll.answers.votes.id']
+			);
+
+			const expected = {
+				poll: {
+					question: '',
+					answers: {
+						items: {
+							answer: '',
+							votes: {
+								items: {
+									id: '',
+								},
+							},
+						},
+					},
+				},
+			};
+
+			expect(selSet).toEqual(expected);
+		});
 	});
 
 	describe('generateSelectionSet', () => {
@@ -301,7 +327,7 @@ describe('flattenItems', () => {
 			const selSet = generateSelectionSet(modelIntroSchema.models, 'Todo');
 
 			const expected =
-				'id name description createdAt updatedAt todoMetaId owner';
+				'id name description status tags createdAt updatedAt todoMetaId owner';
 
 			expect(selSet).toEqual(expected);
 		});