+++ packages/api-graphql/src/internals/server/index.ts
@@ -1,4 +1,4 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-export { generateClient } from './generateClient';
+export { generateClientWithAmplifyInstance } from './generateClientWithAmplifyInstance';