+++ packages/geo/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 3.0.9 (2023-12-22)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
+## 3.0.8 (2023-12-18)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
+## 3.0.7 (2023-12-12)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
+## 3.0.6 (2023-12-05)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
 ## 3.0.5 (2023-11-22)
 
 **Note:** Version bump only for package @aws-amplify/geo